create tx component refactored, new RpcApi.call2() method analyses json rpc error
