+++ src/lib/rpcApi.js
@@ -11,6 +11,29 @@ export class RpcApi {
   }
 
   call(name, params, headers) {
+    return this.jsonPost(name, params, headers);
+  }
+
+  /**
+   * This call analyses JSON RPC response.
+   * It returns promise which resolves whether 'result' field found
+   * or reject in case 'error' field found.
+   *
+   * @returns {Promise}
+   */
+  call2(name, params, headers) {
+    return new Promise((resolve, reject) => {
+      this.jsonPost(name, params, headers).then((json) => {
+        if (json.result)
+          resolve(json.result)
+        else if (json.error)
+          reject(json.error)
+        else new Error(`Unknown JSON RPC response: ${json}`)
+      }).catch(error => reject(error));
+    });
+  }
+
+  jsonPost(name, params, headers) {
     const data = {
       'jsonrpc': '2.0',
       'method': name,
